refactor: separate Legacy and New exportPrivateKey into two files
